Rename monitoring package
